Build External Link Renderer for Portable Text & Reorganize Serializer Components
