+++ packages/web/src/core/components/cms/PortableText/marksComponents/index.ts
@@ -1,0 +1,1 @@
+export * from './ExternalLink';
