+++ packages/web/src/core/components/cms/PortableText/typesComponents/Figure.tsx
@@ -7,6 +7,11 @@ import { clientConfig } from '@app/configs';
 import { Image } from '../../../media';
 import { BodyText } from '../../../typography';
 
+/**
+ * Handles rendering images added to a Sanity Block
+ *
+ * TODO: Improve Types based on what we expect from Sanity
+ */
 export const Figure: React.FC<FigureProps> = ({ node }) => {
     if (!node || !node.asset || !node.asset._id) {
         return null;
@@ -18,9 +23,13 @@ export const Figure: React.FC<FigureProps> = ({ node }) => {
         clientConfig.sanity
     );
 
+    if (!fluidProps) {
+        return null;
+    }
+
     return (
         <figure key={node._key}>
-            <Image fluid={fluidProps!} alt={node.alt} />
+            <Image fluid={fluidProps} alt={node.alt} />
             {node.caption && node.credit && (
                 <BodyText as="figcaption" color="medium" size="s">
                     {node.caption}; photo credit: {node.credit}
