+++ packages/web/src/core/components/cms/PortableText/typesComponents/index.ts
@@ -1,0 +1,2 @@
+export * from './DefaultBlock';
+export * from './Figure';
